requirements.txt
requests
beautifulsoup4
discord.py
python-dateutil
python-dotenv
fastapi
pydantic
uvicorn